refactor(model): 모델의 이름을 파스칼 케이스로 변경
